feat: add webpack config
