+++ package.json
@@ -1,14 +1,14 @@
 {
-  "name": "crud-api",
+  "name": "project",
   "version": "1.0.0",
-  "description": "",
+  "description": "project",
   "main": "server.ts",
   "scripts": {
-    "build": "tsc --build",
-    "clean": "tsc --build --clean",
+    "build:prod": "webpack --mode=production --node-env=production",
+    "build:multi": "cross-env build=multi webpack --mode=production --node-env=production",
     "start:dev": "nodemon ./src/server.ts",
-    "start:prod": "npm run clean && npm run build && node ./dist/server.js",
-    "start:multi": "nodemon ./src/multiserver.ts",
+    "start:prod": "npm run build:prod && node ./dist/server.bundle.js",
+    "start:multi": "npm run build:multi && node ./dist/multiserver.bundle.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "eslint src/**/*.ts"
   },
@@ -22,18 +22,24 @@
     "url": "https://github.com/mashuxa/crud-api/issues"
   },
   "homepage": "https://github.com/mashuxa/crud-api#readme",
-  "dependencies": {
-    "cross-env": "^7.0.3",
-    "dotenv": "^16.0.3",
-    "uuid": "^9.0.0"
-  },
   "devDependencies": {
+    "@babel/core": "^7.20.12",
+    "@babel/preset-env": "^7.20.2",
     "@types/uuid": "^9.0.0",
     "@typescript-eslint/eslint-plugin": "^5.48.0",
     "@typescript-eslint/parser": "^5.48.0",
+    "@webpack-cli/generators": "^3.0.1",
+    "babel-loader": "^9.1.2",
+    "babel-plugin-transform-inline-environment-variables": "^0.4.4",
+    "cross-env": "^7.0.3",
+    "dotenv": "^16.0.3",
     "eslint": "^8.31.0",
     "nodemon": "^2.0.20",
+    "ts-loader": "^9.4.2",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.4"
+    "typescript": "^4.9.4",
+    "uuid": "^9.0.0",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1"
   }
 }
